Use tmux to keep ssh session alive
Using the vagrant ssh command for remote commands comes with the
disadvantage of a lag between command execution and output.

So instead of using the ssh command for every remote command this
commit introduces tmux to keep a ssh session running.

Commands are then fed to the tmux session and the output is piped
back to the calling session.

This approach also keeps the state of the remote machine between
commands (e.g.: a previous ```cd``` will still be honored).